feat(messages): symmetric "Paid" badge — payer + issuer in 1:1, both in groups
PR #451 review: Bolt11 "Paid" badge had visibility gaps:

- 1:1 DMs: only the issuer saw "Paid" reliably — the payer's branch in
  isInvoicePaid checked paidIncomingHashes (wallet tx sync) but NOT the
  persistent settlement cache, so cold starts showed unpaid until tx
  sync re-ran. The cache also only stored issuer-side results.
- Group chats: GroupConversationScreen never passed isInvoicePaid to
  MessageBubble, so the badge was dead code there entirely.

Extracted the tracking logic into src/hooks/usePaidInvoiceTracker.ts:

- Persists BOTH directions to bolt11SettlementCache (per-render diff
  via a ref-tracked seen-set so we only write each hash once per
  session).
- Cold-start hydration loads the cache regardless of direction;
  isInvoicePaid consults `paidHashes` in both branches.
- Issuer-side NWC poll preserved unchanged (15 s, AppState-gated).
  Payer side relies on wallet tx sync — NIP-47 lookupInvoice doesn't
  reliably resolve invoices the wallet didn't issue.

Wired into both ConversationScreen and GroupConversationScreen so
1:1 and group chats use identical settlement logic. Group bubbles
now render Paid for any member whose own wallet records the
counterparty tx.
